[libcu++] Put SFINAE traits into their own file
